+++ Assets/Player/Fatigue.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8fe622d2e47f9594bb1542ad5977b476
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
