+++ Assets/Player/Indicators.cs
@@ -1,0 +1,25 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Indicators : MonoBehaviour
+{
+
+    private bool onWall;
+    private bool onGround;
+    private bool onPullUp;
+    private bool landed;
+
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
