+++ Assets/Player/Indicators.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f84b59a07112e8646a3152f5282a7b45
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
