+++ Assets/Player/PlayerContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3014a29eb470c7f448b4134ba369ec49
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
